+++ DataDefinitions/Organic.cs
@@ -1,44 +1,53 @@
-﻿using Utilities;
+﻿using JetBrains.Annotations;
 using Newtonsoft.Json;
 
 namespace EddiDataDefinitions
 {
     public class Organic
     {
-        [PublicAPI]
+        [ Utilities.PublicAPI ]
         public OrganicGenus genus { get; set; }
 
-        [PublicAPI]
-        public OrganicSpecies species { get; set; }
+        [Utilities.PublicAPI ]
+        public OrganicSpecies species
+        {
+            get => _species;
+            set
+            {
+                _species = value;
+                genus = _species?.genus;
+            }
+        }
 
-        [PublicAPI]
-        public OrganicVariant variant { get; set; }
+        private OrganicSpecies _species;
+
+        [ Utilities.PublicAPI ]
+        public OrganicVariant variant
+        {
+            get => _variant;
+            set
+            {
+                _variant = value;
+                _species = value?.species;
+                genus = value?.genus;
+            }
+        }
+
+        private OrganicVariant _variant;
+
+        [JsonIgnore, Utilities.PublicAPI( "The minimum value from all predictions of this genus." )]
+        public long predictedMinimumValue => valueOverride ?? genusPredictedMinimumValue ?? 0;
+
+        [JsonIgnore, Utilities.PublicAPI( "The maximum value from all predictions of this genus." )]
+        public long predictedMaximumValue => valueOverride ?? genusPredictedMaximumValue ?? 0;
 
         [JsonProperty]
         internal long? genusPredictedMinimumValue = null;
-        public void SetPredictedMinimumValue ( long? minimum )
-        {
-            genusPredictedMinimumValue = minimum;
-        }
 
         [JsonProperty]
         internal long? genusPredictedMaximumValue = null;
-        public void SetPredictedMaximumValue ( long? maximum )
-        {
-            genusPredictedMaximumValue = maximum;
-        }
 
-        [JsonIgnore, PublicAPI( "The minimum value from all predictions of this genus." )]
-        public long predictedMinimumValue => valueOverride ??
-                                             genusPredictedMinimumValue ??
-                                             0;
-
-        [JsonIgnore, PublicAPI( "The maximum value from all predictions of this genus." )]
-        public long predictedMaximumValue => valueOverride ??
-                                             genusPredictedMaximumValue ??
-                                             0;
-
-        [JsonIgnore, PublicAPI( "The minimum distance that you must travel before you can collect a fresh sample of this genus (if known)" )]
+        [JsonIgnore, Utilities.PublicAPI( "The minimum distance that you must travel before you can collect a fresh sample of this genus (if known)" )]
         public int? minimumDistanceMeters => genus?.minimumDistanceMeters;
 
         /// <summary>
@@ -51,59 +60,49 @@ namespace EddiDataDefinitions
         /// </summary>
         //public long? valuePredicted { get; set; }
 
-        [PublicAPI( "The bonus credit value, as awarded when selling organic data" )]
+        [Utilities.PublicAPI( "The bonus credit value, as awarded when selling organic data" )]
         public decimal bonus { get; set; }
-
-        public Organic ()
-        {
-        }
 
         /// <summary>
         /// Populate the organic from variant data. Most preferred.
         /// </summary>
-        public Organic ( OrganicVariant variant )
+        public Organic ( [NotNull] OrganicVariant variant )
         {
-            if ( variant is null ) { return; }
             this.variant = variant;
-            this.species = variant.species;
-            this.genus = variant.species?.genus;
         }
 
         /// <summary>
         /// Populate the organic from species data. Supplement using the {SetVariantData} method when variant data is available.
         /// </summary>
-        public Organic ( OrganicSpecies species )
+        public Organic ( [NotNull] OrganicSpecies species )
         {
-            if ( species is null ) { return; }
             this.species = species;
-            this.genus = species.genus;
         }
 
         /// <summary>
         /// Populate the organic from genus data. Least preferred. Supplement using the {SetVariantData} method when variant data is available.
         /// </summary>
-        public Organic ( OrganicGenus genus )
+        public Organic ( [NotNull] OrganicGenus genus )
         {
-            if ( genus is null ) { return; }
             this.genus = genus;
         }
 
+        public void SetPredictedMinimumValue ( long? minimum )
+        {
+            genusPredictedMinimumValue = minimum;
+        }
+
+        public void SetPredictedMaximumValue ( long? maximum )
+        {
+            genusPredictedMaximumValue = maximum;
+        }
+
         /// <summary> Get all the biological data, this should be done at the first sample </summary>
-        [PublicAPI]
+        [Utilities.PublicAPI]
         public static Organic Lookup ( long entryid, string variant )
         {
             var organicVariant = OrganicVariant.Lookup( entryid, variant );
-            return new Organic( organicVariant );
-        }
-
-        /// <summary>Get all the biological data, this should be done at the first sample</summary>
-        [PublicAPI]
-        public void SetVariantData ( OrganicVariant thisVariant )
-        {
-            if ( thisVariant is null ) { return; }
-            this.variant = thisVariant;
-            this.species = this.variant?.species;
-            this.genus = this.variant?.species?.genus;
+            return organicVariant is null ? null : new Organic( organicVariant );
         }
     }
 }